SINGA-349 Create layer operations for autograd
1. modified the operation: Linear

2. reassemble codes: move Operation, Dummy from tensor.py to autograd.py
